pyproject.toml
[tool.poetry]
name = "diabetes_classifier"
version = "0.0.1"
description = ""
authors = ["John Cummings <jcummingsutk@gmail.com>"]
readme = "README.md"
packages = [{ include = "diabetes_classifier" }]

[tool.poetry.dependencies]
diabetes-classifier-models = { path = "diabetes_classifier/diabetes_classifier_models", develop = true }


python = ">=3.10, <3.12"
dvc = "^3.32.0"
dvc-s3 = "^2.23.0"
pandas = "^2.1.3"
numpy = "^1.26.2"
azure-ai-ml = "^1.12.1"
azure-identity = "^1.15.0"
imbalanced-learn = "^0.11.0"
plotly = "^5.18.0"
nbformat = "^5.9.2"
dvc-azure = "^3.0.1"
python-dotenv = "^1.0.0"
seaborn = "^0.13.0"
azureml-mlflow = "^1.54.0.post1"
azureml-core = "^1.54.0.post1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.11.0"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
ipykernel = "^6.27.1"

[tool.poetry.scripts]
train_xgboost_model = "diabetes_classifier.models.xgboost.train:main"

[tool.poetry.group.model.dependencies]
mlflow = "^2.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.flake8]
max-line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = ["."]
